Try out a deletable list item.
I think making a general-use thing like this is the right abstraction. It seems to
work pretty transparently with other stylings, which is nice. Note that I have to
use my fork of the project to get some custom behaviors. Next up: figuring out which
recipes are deletable and how to delete them.