Fix bug where focus() would throw error when you edited multiple recipes in one session.
The bug was that _.memoize was being called at class definition time and wasn't accounting
for the this-binding of the particular instance of the class, so another instance of the
component would get the same memoized function as the first instance to call it with that
parameter. Instead, bind in componentWillMount so it's per-instance.

The impetus for memoizing is to reduce prop churn, though I guess it doesn't really matter
since the inputs are being rerendered a whole bunch of times anyway. Whatever.