Decaffeinate to remove all coffeescript!
The following script was used:

mkdir -p /tmp/transpile

for f in **/*.js;
do
  echo "$f"
  tmpname="/tmp/transpile/$f"
  mkdir -p $(dirname "$tmpname")
  decaffeinate < "$f" > "$tmpname" --keep-commonjs --prefer-const --loose-default-params
done

cp -r /tmp/transpile/* .

It appears that decaffeinate had some struggles with its tree traversal, and
would produce empty files in many uses. Explicitly passing one file at a time
and saving off the output to a file that was NOT the input file prevented
this issue and is what the above loop does.