Strip base liquor from the code entirely.
While a neat feature, it doesn't make sense in the new version of
search. Once liqueurs have a broad category, there will be nothing
novel this _data_ can provide. The UI was neat, and much more fluid
that the current workaround (remove old base -> search new base ->
add new base -> rinse, repeat), but that behavior can be implemented
with the data as-is.